chore: fix build on refactoring
