+++ packages/core/src/Element.ts
@@ -1,4 +1,4 @@
-import { PlaywrightClient, launch } from './driver/Playwright'
+import { launch } from './driver/Playwright'
 import { IRunner, Runner, PersistentRunner } from './Runner'
 import { mustCompileFile } from './TestScript'
 import { TestScriptOptions } from './TestScriptOptions'
@@ -7,10 +7,10 @@ import { ElementOptions, ElementRunArguments, normalizeElementOptions } from './
 import { CustomConsole, ReportCache } from '@flood/element-report'
 import chalk from 'chalk'
 import { EventEmitter } from 'events'
-import { TestSettings } from './runtime/Settings'
-import { TestObserver } from './runtime/test-observers/TestObserver'
-import { AsyncFactory } from './utils/Factory'
-import { BROWSER_TYPE } from './page/types'
+// import { TestSettings } from './runtime/Settings'
+// import { TestObserver } from './runtime/test-observers/TestObserver'
+// import { AsyncFactory } from './utils/Factory'
+// import { BROWSER_TYPE } from './page/types'
 
 async function runSingleTestScript(opts: ElementOptions): Promise<void> {
 	const { testScript, clientFactory } = opts
@@ -67,6 +67,10 @@ async function runSingleTestScript(opts: ElementOptions): Promise<void> {
 		// eslint-disable-next-line no-ex-assign
 		err = null
 	}
+}
+
+export async function runCommandLineOrigin(args: ElementOptions) {
+	//do sth
 }
 
 export async function runCommandLine(args: ElementRunArguments): Promise<void> {
