+++ packages/core/src/runtime/test-observers/Context.ts
@@ -21,7 +21,7 @@ export class Context {
 	public async attachToPage(reporter: IReporter, page: Page) {
 		this.networkRecorder = new NetworkRecorder(page)
 		await this.networkRecorder.attachEvents()
-		this.observer = new NetworkObserver(reporter, this.networkRecorder)
+		this.observer = new NetworkObserver(reporter, this.networkRecorder, [])
 		await this.observer.attachToNetworkRecorder()
 	}
