+++ packages/core/tsconfig.json
@@ -1,12 +1,22 @@
 {
-	"extends": "../../tsconfig.json",
-	"compilerOptions": {
-		"declaration": true,
-		"outDir": "dist"
-	},
-
-	"references": [{ "path": "../compiler" }],
-	"include": ["**/*.ts"],
-
-	"files": ["./index.ts", "./api.ts"]
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "declaration": true,
+    "outDir": "dist"
+  },
+  "references": [
+    {
+      "path": "../compiler"
+    },
+    {
+      "path": "../report"
+    }
+  ],
+  "include": [
+    "**/*.ts"
+  ],
+  "files": [
+    "./index.ts",
+    "./api.ts"
+  ]
 }
