+++ packages/element-api/package.json
@@ -9,7 +9,8 @@
 		"build": "tsc -b"
 	},
 	"dependencies": {
-		"@flood/element-core": "^2.0.0-beta.4"
+		"@flood/element-core": "^2.0.0-beta.4",
+		"@flood/element-report": "^2.0.0-beta.4"
 	},
 	"devDependencies": {
 		"@types/node": "12.7.4"
