+++ packages/flood-runner/package.json
@@ -24,7 +24,7 @@
 	"dependencies": {
 		"@flood/element-core": "^2.0.0-beta.4",
 		"@flood/node-influx": "^5.0.9",
-		"@flood/element-report": "^1.3.0-beta.19",
+		"@flood/element-report": "^2.0.0-beta.4",
 		"find-root": "^1.1.0",
 		"winston": "^3.2.1"
 	}
