+++ packages/flood-runner/src/Grid.ts
@@ -1,4 +1,4 @@
-import { runCommandLine, ElementOptions, TestObserver } from '@flood/element-api'
+import { runCommandLineOrigin, ElementOptions, TestObserver } from '@flood/element-api'
 import { Context, TimingObserver, BROWSER_TYPE } from '@flood/element-core'
 import { TracingObserver } from './test-observers/Tracing'
 import { initConfig } from './initConfig'
@@ -43,5 +43,5 @@ export async function run(file: string): Promise<void> {
 
 	startConcurrencyTicker(influxReporter)
 
-	return runCommandLine(opts)
+	return runCommandLineOrigin(opts)
 }
