+++ packages/report/package.json
@@ -1,6 +1,6 @@
 {
 	"name": "@flood/element-report",
-	"version": "1.3.0-beta.19",
+	"version": "2.0.0-beta.4",
 	"description": "Report for Element",
 	"author": "Ivan Vanderbyl <ivanvanderbyl@gmail.com>",
 	"homepage": "https://github.com/flood-io/element#readme",
